feat: Integrating CASL

+++ src/components/users/users.controller.ts
@@ -13,9 +13,9 @@ import {
 } from '@nestjs/common';
 import { ApiBearerAuth, ApiTags } from '@nestjs/swagger';
 import { FilterQuery } from 'mongoose';
-import { Authorization } from 'src/common/authorization';
-import { Roles } from 'src/common/constant';
-import { ValidationError } from 'src/common/errors';
+import { CaslAbilityFactory } from 'src/casl/casl-ability.factory';
+import { Action } from 'src/common/constant';
+import { UnauthorizedError, ValidationError } from 'src/common/errors';
 import { ROUTES } from 'src/common/routes';
 import { capitalizeInitials } from 'src/utils';
 import { JwtAuthGuard } from '../auth/jwt-auth.guard';
@@ -27,7 +27,10 @@ import { validateUser } from './users.validation';
 @ApiTags(capitalizeInitials(ROUTES.users.name))
 @Controller(ROUTES.users.route)
 export class UsersController {
-  constructor(private readonly userService: UsersService) {}
+  constructor(
+    private readonly userService: UsersService,
+    private caslAbilityFactory: CaslAbilityFactory,
+  ) {}
 
   @Post()
   async create(@Body() userDto: UsersDto): Promise<User> {
@@ -43,13 +46,23 @@ export class UsersController {
   @UseGuards(JwtAuthGuard)
   @Put('log-out')
   async logOut(@Request() req: any): Promise<any> {
+    const ability = this.caslAbilityFactory.createForUser(req.user);
+    if (!ability.can(Action.Update, User)) {
+      throw new UnauthorizedError();
+    }
+
     return await this.userService.logOut(req.user);
   }
 
   @ApiBearerAuth()
   @UseGuards(JwtAuthGuard)
   @Get(':id')
-  async get(@Param('id') userId: string): Promise<User> {
+  async get(@Request() req, @Param('id') userId: string): Promise<User> {
+    const ability = this.caslAbilityFactory.createForUser(req.user);
+    if (!ability.can(Action.Read, User)) {
+      throw new UnauthorizedError();
+    }
+
     const conditions: FilterQuery<User> = { _id: userId };
     return await this.userService.findOne(conditions);
   }
@@ -59,7 +72,11 @@ export class UsersController {
   @HttpCode(HttpStatus.NO_CONTENT)
   @Delete(':id')
   async delete(@Request() req, @Param('id') userId: string): Promise<any> {
-    Authorization.authorize(req.user, [Roles.SUPER_ADMIN, Roles.ADMIN]);
+    const ability = this.caslAbilityFactory.createForUser(req.user);
+    if (!ability.can(Action.Delete, User)) {
+      throw new UnauthorizedError();
+    }
+
     return await this.userService.delete(userId);
   }
 }
